bugs

+++ src/bounding_box_model/spatial_bb/components.py
@@ -41,9 +41,8 @@ class SpatialMappingCNN(nn.Module):
         # ROTATE
         # ---------------
         b = x[:, 4, ...]
-        b = b.float()
         #import pdb; pdb.set_trace()
-        b = torch.rot90(b, 1, [2, 3])
+        b = torch.rot90(b.half(), 1, [2, 3])
         #b = b.type_as(x[0])
         b = F.relu(self.b_conv(b))
